third_party/WebKit/LayoutTests/http/tests/inspector/security/security-secure-but-malicious-expected.txt

Tests summary and explanation when all page resources are transferred securely (no-mixed-content) but page is flagged by Google Safe Browsing.

<DIV class=security-summary-text >
Test: Summary Override Text
</DIV>
<DIV class=security-explanation security-explanation-secure >
    <DIV class=security-property security-property-secure >
    </DIV>
    <DIV class=security-explanation-text >
        <DIV class=security-explanation-title >
Secure Resources
        </DIV>
        <DIV >
All resources on this page are served securely.
        </DIV>
    </DIV>
</DIV>
